Show specs on reload and render newSpecs after oldSpecs are deleted.
The Spec edit view was broken in two place. The saved specs weren't
being shown on reload and when the user changed category, the old
specs weren't being deleted. To fix this I had to do some asynchronus
control flow using async.js's each method which provides a callback
*after* a group of methods finish. The old destroy code
`this.collection.invoke('destroy`);` was getting interupted by the
rendering of the new specs. I also refactored the `switchCategory`
method into a `render` method to allow for a callback post oldSpec
deletion. Making saved specs show up on reload required the
initialize method funtionality to be run *after* a fetch followed
by a merge of the saved specs with the current category's
spec collection.